+++ client/apps/game/src/ui/components/trading/market-order-panel.tsx
@@ -2,22 +2,17 @@ import { ConfirmationPopup } from "@/ui/components/bank/confirmation-popup";
 import Button from "@/ui/elements/button";
 import { NumberInput } from "@/ui/elements/number-input";
 import { ResourceIcon } from "@/ui/elements/resource-icon";
+import { currencyFormat, divideByPrecision, formatNumber, multiplyByPrecision } from "@/ui/utils/utils";
 import {
   calculateDonkeysNeeded,
-  currencyFormat,
-  divideByPrecision,
-  formatNumber,
-  multiplyByPrecision,
-} from "@/ui/utils/utils";
-import {
-  DONKEY_ENTITY_TYPE,
-  ONE_MONTH,
-  ResourceManager,
-  ResourcesIds,
   configManager,
+  DONKEY_ENTITY_TYPE,
   findResourceById,
   getRealmAddressName,
   getTotalResourceWeight,
+  ONE_MONTH,
+  ResourceManager,
+  ResourcesIds,
   type ID,
   type MarketInterface,
 } from "@bibliothecadao/eternum";
